pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = ""
authors = [
    {name = "marha-hwang",email = "hwangharam@naver.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.121.1,<0.122.0",
    "uvicorn>=0.20.0,<1.0.0",
    "python-jose>=3.5.0,<4.0.0",
    "python-multipart>=0.0.5,<0.1.0",
    "passlib==1.7.4",
    "bcrypt==4.0.1",
    "sqlalchemy>=2.0.44,<3.0.0",
    "pymysql>=1.1.2,<2.0.0",
    "openai>=2.8.1,<3.0.0",
    "pydantic-settings>=2.12.0",
    "langchain>=0.3.0,<0.4.0",
    "langchain-google-genai>=2.1.2,<3.0.0",
    "langchain-core>=0.3.10,<0.4.0",
    "langchain-community>=0.3.0,<0.4.0",
    "langchain-openai>=0.3.12,<0.4.0",
    "langchain-anthropic>=0.3.10,<0.4.0",
    "langchain-weaviate>=0.0.3,<0.1.0",
    "langgraph>=0.4.5",
    "beautifulsoup4>=4.12.2,<5.0.0",
    "weaviate-client>=4.0.0,<5.0.0",
    "lxml>=6,<7.0.0",
    "voyageai>=0.1.4,<0.2.0",
    "pillow>=10.2.0,<11.0.0",
    "psycopg2-binary>=2.9.9,<3.0.0",
    "langgraph-sdk>=0.2,<0.3",
    "pytest>=7.3.0,<8.0.0",
    "ruff>=0.13.3,<0.14.0",
    "pandas>=2.2.2,<3.0.0",
    "ipykernel>=7.1.0",
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "python-dotenv>=1.2.1",
    "langfuse>=3.11.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
